WS-47: Initial featured post component. Move parseLikes to parses.ts for exporting.
