+++ frontend/src/components/Post/CardPost.tsx
@@ -1,22 +1,10 @@
 import { PostData } from "../../pages/Posts";
+import { parseLikes } from "../../util/parses";
 import "./CardPost.scss";
 
 interface PostDataProps {
   postData: PostData;
 }
-
-const parseLikes = (likes: number) => {
-  if (likes > 1000) {
-    const thousands = Math.floor(likes / 1000);
-    const hundreds = Math.floor((likes / 1000 - thousands) * 10);
-    if (hundreds === 0) {
-      return `${thousands} k`;
-    } else {
-      return `${thousands}.${hundreds} k`;
-    }
-  }
-  return likes.toString();
-};
 
 const CardPost = ({ postData }: PostDataProps) => {
   return (
