+++ frontend/src/components/Post/FeaturedPost.scss
@@ -1,0 +1,2 @@
+@use "../../styles/variables" as *;
+@use "../../styles/mixins" as *;
